Makes MORE things call the attack_self() parent
Yes, I could make special_handling a var on obj/item HOWEVER i want it to be specific so it can be tracked down later and ONLY the objects that use it can be refactored instead of sitting there literally forever and it just becoming 'a thing'.

+++ code/game/objects/items/weapons/mop_deploy.dm
@@ -55,6 +55,9 @@
 	. = ..()
 
 /obj/item/mop_deploy/attack_self(mob/user as mob)
+	. = ..()
+	if(.)
+		return TRUE
 	user.drop_from_inventory(src)
 	spawn(1) if(!QDELETED(src)) qdel(src)
