+++ code/game/objects/structures/stool_bed_chair_nest/bed.dm
@@ -264,6 +264,9 @@
 	pickup_sound = 'sound/items/pickup/axe.ogg'
 
 /obj/item/roller/attack_self(mob/user)
+	. = ..()
+	if(.)
+		return TRUE
 	var/obj/structure/bed/roller/R = new bedtype(user.loc)
 	R.add_fingerprint(user)
 	qdel(src)
@@ -300,7 +303,9 @@
 	held = new /obj/item/roller(src)
 
 /obj/item/roller_holder/attack_self(mob/user as mob)
-
+	. = ..()
+	if(.)
+		return TRUE
 	if(!held)
 		to_chat(user, span_notice("The rack is empty."))
 		return
